requirements.txt
django<3
djangorestframework==3.9.0
djangorestframework-camel-case==1.0.3
django-extra-fields
drf_yasg
Pillow
django-jet
python_http_client